TrainEventList class now removes duplicated indices
Now when an instance of class TrainEventList writes a list of events tried for training, it filters out duplicates. Such a filter is
needed when two samples contain the same source file. Despite in any reasonalble application event selections for such samples must be
pairwise orthogonal, an event to be tried for training is picked up before the event selection is evaluated (which allows to deal with
cases when the selection introduces a bias). It means that the same event can be chosen during processing of each sample, and it would be duplicated in the list if not for the filtering.
ignored, and an empty message aborts the commit. # On branch master # Changes to be committed: # (use "git reset HEAD <file>..." to
unstage) # # modified:  bnn-hep/include/TrainEventList.hpp # modified:  bnn-hep/src/TrainEventList.cpp #